Move diff_string models to separate package
