+++ models/diff_string/char_cnn/README.md
