+++ models/diff_string/char_cnn/model.py
