+++ models/diff_string/kim_cnn/README.md
