+++ models/diff_string/kim_cnn/__main__.py
