+++ models/diff_string/kim_cnn/model.py
