+++ models/diff_string/reg_lstm/README.md
