+++ models/diff_string/reg_lstm/model.py
